[Feature] Add Media Connector SDK methods
