+++ src/tests/__mocks__/MockEditorAPI.ts
@@ -1,0 +1,212 @@
+export const mockAddLayout = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockRemoveLayout = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSetLayoutName = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSelectLayout = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockDuplicateLayout = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockResetLayout = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetFrames = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetSelectedFrames = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetFramesByPageId = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetFrameByName = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetFrameById = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetFramePropertiesOnSelectedLayout = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetFramePropertiesByFrameId = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetFramesProperties = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSelectFrames = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSetFrameHeight = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSetFrameWidth = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSetFrameRotation = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSetFrameVisibility = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockRemoveFrame = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSetFrameX = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSetFrameY = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockResetFrame = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockResetFrameX = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockResetFrameY = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockResetFrameWidth = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockResetFrameHeight = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockResetFrameRotation = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSetFrameAniation = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockTogglePlaybackAnimation = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockResetFrameSize = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetLayouts = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetLayoutById = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetLayoutByName = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetSelectedLayout = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mocksetLayoutHeight = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mocksetLayoutWidth = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockResetLayoutHeight = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockResetLayoutWidth = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetCurrentDocumentState = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockLoadDocument = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockAddVariable = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockRemoveVariables = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGroupVariables = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSetVariableName = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSetVariableLabel = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSetVariableType = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSetDefaultVariableValue = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSetVariableValue = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockMoveVariable = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockMoveVariables = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSetVariableIsHidden = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSetVariableIsRequired = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSetVariableIsReadonly = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockUngroupVariable = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockDuplicateVariable = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetVariables = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetVariableById = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetVariableByName = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetSelectedTool = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSetTool = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockPlayAnimation = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockPauseAnimation = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSetScrubberPosition = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockSetAnimationDuration = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockResetFrameAnimation = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockResetAnimation = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetPages = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetPageById = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetAnimationsOnSelectedLayout = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetAnimationByFrameId = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetAnimationsByLayoutId = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetLogs = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockToggleDebugPanel = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockEnableDebug = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockDisableDebug = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockUndo = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockRedo = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockDeltaUpdate = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetColors = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetColorById = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockCreateColor = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockDuplicateColor = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockMoveColors = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockRenameColor = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockUpdateColor = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockRemoveColor = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetParagraphStyles = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockGetParagraphStyleById = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockCreateParagraphStyle = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockDuplicateParagraphStyle = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockRenameParagraphStyle = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockRemoveParagraphStyle = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockMoveParagraphStyles = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockUpdateParagraphStyle = jest.fn().mockResolvedValue({ success: true, status: 0 });
+export const mockMediaConnectorQuery = jest.fn().mockResolvedValue({success: true, status: 0});
+export const mockMediaConnectorDownload = jest.fn().mockResolvedValue({success: true, status: 0});
+export const mockMediaConnectorRemove = jest.fn().mockResolvedValue({success: true, status: 0});
+export const mockMediaConnectorCopy = jest.fn().mockResolvedValue({success: true, status: 0});
+export const mockMediaConnectorUpload = jest.fn().mockResolvedValue({success: true, status: 0});
+export const mockMediaConnectorGetCapabilities = jest.fn().mockResolvedValue({success: true, status: 0});
+export const mockMediaConnectorGetQueryOptions = jest.fn().mockResolvedValue({success: true, status: 0});
+export const mockMediaConnectorGetDownloadOptions = jest.fn().mockResolvedValue({success: true, status: 0});
+
+const MockEditorAPI = {
+    addLayout: mockAddLayout,
+    removeLayout: mockRemoveLayout,
+    renameLayout: mockSetLayoutName,
+    renameFrame: mockSetLayoutName,
+    selectLayout: mockSelectLayout,
+    duplicateLayout: mockDuplicateLayout,
+    resetLayout: mockResetLayout,
+    getFrames: mockGetFrames,
+    getSelectedFrames: mockGetSelectedFrames,
+    getFramesByPageId: mockGetFramesByPageId,
+    getFrameByName: mockGetFrameByName,
+    getFrameById: mockGetFrameById,
+    getFramePropertiesOnSelectedLayout: mockGetFramePropertiesOnSelectedLayout,
+    getFramePropertiesByFrameId: mockGetFramePropertiesByFrameId,
+    getFramesProperties: mockGetFramesProperties,
+    selectFrames: mockSelectFrames,
+    setFrameHeight: mockSetFrameHeight,
+    setFrameWidth: mockSetFrameWidth,
+    setFrameVisibility: mockSetFrameVisibility,
+    removeFrame: mockRemoveFrame,
+    setFrameX: mockSetFrameX,
+    setFrameY: mockSetFrameY,
+    setFrameRotation: mockSetFrameRotation,
+    resetFrame: mockResetFrame,
+    resetFrameX: mockResetFrameX,
+    resetFrameY: mockResetFrameY,
+    resetFrameWidth: mockResetFrameWidth,
+    resetFrameHeight: mockResetFrameHeight,
+    resetFrameRotation: mockResetFrameRotation,
+    setFrameAnimation: mockSetFrameAniation,
+    getAnimationsOnSelectedLayout: mockGetAnimationsOnSelectedLayout,
+    getAnimationByFrameId: mockGetAnimationByFrameId,
+    getAnimationsByLayoutId: mockGetAnimationsByLayoutId,
+    playAnimation: mockPlayAnimation,
+    pauseAnimation: mockPauseAnimation,
+    setScrubberPosition: mockSetScrubberPosition,
+    setAnimationDuration: mockSetAnimationDuration,
+    resetFrameAnimation: mockResetFrameAnimation,
+    resetAnimation: mockResetAnimation,
+    togglePlaybackAnimation: mockTogglePlaybackAnimation,
+    resetFrameSize: mockResetFrameSize,
+    getLayouts: mockGetLayouts,
+    getLayoutById: mockGetLayoutById,
+    getLayoutByName: mockGetLayoutByName,
+    getSelectedLayout: mockGetSelectedLayout,
+    setLayoutHeight: mocksetLayoutHeight,
+    setLayoutWidth: mocksetLayoutWidth,
+    resetLayoutHeight: mockResetLayoutHeight,
+    resetLayoutWidth: mockResetLayoutWidth,
+    getCurrentDocumentState: mockGetCurrentDocumentState,
+    loadDocument: mockLoadDocument,
+    getVariables: mockGetVariables,
+    getVariableById: mockGetVariableById,
+    getVariableByName: mockGetVariableByName,
+    addVariable: mockAddVariable,
+    removeVariables: mockRemoveVariables,
+    groupVariables: mockGroupVariables,
+    setVariableName: mockSetVariableName,
+    setVariableLabel: mockSetVariableLabel,
+    setVariableType: mockSetVariableType,
+    setDefaultVariableValue: mockSetDefaultVariableValue,
+    setVariableValue: mockSetVariableValue,
+    duplicateVariable: mockDuplicateVariable,
+    moveVariable: mockMoveVariable,
+    moveVariables: mockMoveVariables,
+    setVariableIsHidden: mockSetVariableIsHidden,
+    setVariableIsRequired: mockSetVariableIsRequired,
+    setVariableIsReadonly: mockSetVariableIsReadonly,
+    ungroupVariable: mockUngroupVariable,
+    getSelectedTool: mockGetSelectedTool,
+    setTool: mockSetTool,
+    getPages: mockGetPages,
+    getPageById: mockGetPageById,
+    getLogs: mockGetLogs,
+    enableDebug: mockEnableDebug,
+    disableDebug: mockDisableDebug,
+    toggleDebugPanel: mockToggleDebugPanel,
+    undo: mockUndo,
+    redo: mockRedo,
+    selectedTextStyleDeltaUpdate: mockDeltaUpdate,
+    getColors: mockGetColors,
+    getColorById: mockGetColorById,
+    createColor: mockCreateColor,
+    duplicateColor: mockDuplicateColor,
+    moveColors: mockMoveColors,
+    renameColor: mockRenameColor,
+    updateColor: mockUpdateColor,
+    removeColor: mockRemoveColor,
+    getParagraphStyles: mockGetParagraphStyles,
+    getParagraphStyleById: mockGetParagraphStyleById,
+    createParagraphStyle: mockCreateParagraphStyle,
+    duplicateParagraphStyle: mockDuplicateParagraphStyle,
+    renameParagraphStyle: mockRenameParagraphStyle,
+    removeParagraphStyle: mockRemoveParagraphStyle,
+    moveParagraphStyles: mockMoveParagraphStyles,
+    updateParagraphStyle: mockUpdateParagraphStyle,
+    mediaConnectorCopy: mockMediaConnectorCopy,
+    mediaConnectorQuery: mockMediaConnectorQuery,
+    mediaConnectorDownload: mockMediaConnectorDownload,
+    mediaConnectorRemove: mockMediaConnectorRemove,
+    mediaConnectorUpload: mockMediaConnectorUpload,
+    mediaConnectorGetCapabilities: mockMediaConnectorGetCapabilities,
+    mediaConnectorGetQueryOptions: mockMediaConnectorGetQueryOptions,
+    mediaConnectorGetDownloadOptions: mockMediaConnectorGetDownloadOptions,
+};
+
+export default MockEditorAPI;
