+++ src/tests/controllers/PageController.test.ts
@@ -1,5 +1,5 @@
-import MockEditorAPI from '../__mocks__/FrameProperties';
 import { PageController } from '../../controllers/PageController';
+import MockEditorAPI from "../__mocks__/MockEditorAPI";
 
 let mockedPageProperties: PageController;
