+++ src/tests/controllers/ParagraphStyleController.test.ts
@@ -1,8 +1,8 @@
-import MockEditorAPI from '../__mocks__/FrameProperties';
 import { ColorType } from '../../../types/ColorStyleTypes';
 import { ParagraphStyleController } from '../../controllers/ParagraphStyleController';
 import { ColorUsageType, ParagraphStyleUpdate } from '../../../types/ParagraphStyleTypes';
 import { Alignment, Case, Scripting } from '../../../types/TextStyleTypes';
+import MockEditorAPI from "../__mocks__/MockEditorAPI";
 
 let mockedParagraphStyleController: ParagraphStyleController;
