+++ src/tests/index.test.ts
@@ -1,6 +1,6 @@
 import { SDK } from '../index';
 import mockConfig from './__mocks__/config';
-import MockEditorAPI from './__mocks__/FrameProperties';
+import MockEditorAPI from "./__mocks__/MockEditorAPI";
 
 let mockedSDK: SDK;
