+++ src/utils/EditorResponseData.ts
@@ -11,3 +11,13 @@ export function getEditorResponseData<T>(response: EditorResponse<unknown>, pars
                 : null,
     };
 }
+
+/**
+ * This method should be used with care. It allows marshalling raw data
+ * to and from the studio engine. This is usefull to avoid serialization
+ * of e.g. byte arrays
+ * @param response
+ */
+export function getRawEditorResponseData<T>(response: T): T {
+    return response;
+}
